requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-telegram-bot==20.7
python-dotenv==1.0.1
schedule==1.2.1
lxml==5.1.0
